requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
gunicorn==20.1.0
importlib-metadata==4.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0